Cleaning up “common” and some dead code.
- Most of it goes into account-settings for now.
- Shuffling the “utils” files around to classify them better.
- Removing unused assets
