+++ src/account-settings/AccountSettingsPage.jsx
@@ -15,7 +15,8 @@ import { Hyperlink } from '@edx/paragon';
 import messages from './AccountSettingsPage.messages';
 import { fetchSettings, saveSettings, updateDraft } from './actions';
 import { accountSettingsPageSelector } from './selectors';
-import { Alert, PageLoading } from '../common';
+import PageLoading from './PageLoading';
+import Alert from './Alert';
 import JumpNav from './JumpNav';
 import DeleteAccount from './delete-account';
 import EditableField from './EditableField';
