+++ src/account-settings/EmailField.jsx
@@ -6,7 +6,8 @@ import { Button, StatefulButton, Input, ValidationFormGroup } from '@edx/paragon
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faExclamationTriangle, faPencilAlt } from '@fortawesome/free-solid-svg-icons';
 
-import { Alert, SwitchContent } from '../common';
+import Alert from './Alert';
+import SwitchContent from './SwitchContent';
 import messages from './AccountSettingsPage.messages';
 
 import {
