--- src/common/index.js
@@ -1,13 +1,0 @@
-import * as utils from './utils';
-import Alert from './components/Alert';
-import PageLoading from './components/PageLoading';
-import ErrorBoundary from './components/ErrorBoundary';
-import SwitchContent from './components/SwitchContent';
-
-export {
-  Alert,
-  ErrorBoundary,
-  PageLoading,
-  SwitchContent,
-  utils,
-};
